Add tests for Player and PlayerMark objects
`Player` will be a reference to a participant in the game, containing
the player's label ("X" or "O"), and a PlayerMark object.
It will also handle events from the canvas (if local) or the network (if
remote), and passing them along to the gamestate for validation and
recording.

`PlayerMark` stores a path to use as the player's piece in the game, and
rendering options.